Added --post-lex-unfiltered option to lark_parse.py and fixed parsing bug.
Added output of tokens from post-lexer for debugging purposes.  A
post-lex filter was added to take these back out, since they are
not required.  This was done to implement the new command-line
option that allows these to be printed to the screen for
debugging purposes.

The parsing results for the test files are now the same as for
the parglare parser.  The bug was in post_lex.py and had to do
with parsing of partial unindents.